+++ tests/testthat/test-argparse.R
@@ -42,7 +42,7 @@ test_that("print_help works as expected", {
 context("convert_agument")
 test_that("convert_argument works as expected", {
     expect_equal(convert_argument("foobar"), "'foobar'")
-    expect_equal(convert_argument(14.9), 14.9)
+    expect_equal(convert_argument(14.9), "14.9")
     expect_equal(convert_argument(c(12.1, 14.9)), "(12.1, 14.9)")
     expect_equal(convert_argument(c("a", "b")), "('a', 'b')")
 })
@@ -142,6 +142,12 @@ test_that("ArgumentParser works as expected", {
     expect_output(parser$print_help(), "foobar arg1 arg2")
     expect_output(parser$print_help(), "foobar's saying \\(default: bye\\)")
     expect_error(ArgumentParser(python_cmd="foobar"))
+    if (interactive()) {
+        # Bug report by George Chlipala
+        expect_error(ArgumentParser()$parse_args("-h"), "help requested")
+        expect_error(ArgumentParser(add_help=TRUE)$parse_args("-h"), "help requested")
+        expect_error(ArgumentParser(add_help=FALSE)$parse_args("-h"), "unrecognized arguments")
+    }
 })
 test_that("parse_args works as expected", {
     parser <- ArgumentParser(prog="foobar", usage="%(prog)s arg1 arg2")